requirements.txt
numpy
fastapi
uvicorn
obspy
scipy
pymysql
python-dotenv